+++ src/settings/include/shortcuts.h
@@ -95,6 +95,8 @@ struct ShortcutAction {
     static constexpr auto LogViewJumpToRightOfScreen = "logview.jump_to_right";
     static constexpr auto LogViewJumpToLine = "logview.jump_to_line";
     static constexpr auto LogViewJumpToTimestamp = "logview.jump_to_timestamp";
+    static constexpr auto LogViewSearchLimitsTimeRange = "logview.search_limits_time_range";
+    static constexpr auto LogViewSearchLimitsAroundLine = "logview.search_limits_around_line";
     static constexpr auto LogViewJumpToLineNumber = "logview.jump_to_line_number";
 
     static constexpr auto LogViewQfSelectedForward = "logview.qf_selected_forward";