+++ src/settings/src/shortcuts.cpp
@@ -507,6 +507,20 @@ const ShortcutAction::ShortcutList& ShortcutAction::defaultShortcutList()
             },
         },
         {
+            LogViewSearchLimitsTimeRange,
+            {
+                QApplication::tr( "Set search limits to a time range" ),
+                QStringList{ "Ctrl+Alt+T" },
+            },
+        },
+        {
+            LogViewSearchLimitsAroundLine,
+            {
+                QApplication::tr( "Set search limits around the current line" ),
+                QStringList{ "Ctrl+Alt+W" },
+            },
+        },
+        {
             LogViewQfForward,
             {
                 QApplication::tr( "Main view: find next" ),