+++ src/ui/src/mainwindowtext.cpp
@@ -47,6 +47,13 @@ const char* action::goToLineStatusTip
 const char* action::goToTimestampText = QT_TR_NOOP( "Go to timestamp..." );
 const char* action::goToTimestampStatusTip
     = QT_TR_NOOP( "Scrolls selected main view to the first line at or after a time" );
+const char* action::searchLimitsTimeRangeText = QT_TR_NOOP( "Set search limits to time range..." );
+const char* action::searchLimitsTimeRangeStatusTip
+    = QT_TR_NOOP( "Limits the search to the log lines between a start and an end time" );
+const char* action::searchLimitsAroundLineText
+    = QT_TR_NOOP( "Set search limits around current line..." );
+const char* action::searchLimitsAroundLineStatusTip
+    = QT_TR_NOOP( "Limits the search to N minutes before and after the current log line" );
 const char* action::findText = QT_TR_NOOP( "&Find..." );
 const char* action::findStatusTip = QT_TR_NOOP( "Find the text" );
 const char* action::clearLogText = QT_TR_NOOP( "Clear file..." );